+++ lib/rails/generators/neo4j/install/templates/neo4j_rails5.rb
@@ -1,0 +1,10 @@
+# Configure NEO4J as the ORM
+Rails.application.config.generators do |g|
+  g.orm             :neo4j
+end
+
+# Configure where the neo4j database should exist
+Neo4j::Config.use do |config|
+  config[:storage_path] = Rails.application.config.root.join("db/neo4j-#{Rails.env}")
+  #"#{Rails.application.config.root}/db/neo4j-#{Rails.env}"
+end